Update to treat compiler warnings as errors, and to restructure Forum.Data package code
