+++ Forum.sln
@@ -9,9 +9,11 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Forum.Data", "src\Forum.Dat
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "test", "test", "{E96FD598-6925-4A2B-B1A1-5B03AAB9C1B7}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Data.Tests", "test\Data.Tests\Data.Tests.csproj", "{1CDA37F8-9EF1-4349-8473-E529DC09482D}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Forum.Data.Tests", "test\Forum.Data.Tests\Forum.Data.Tests.csproj", "{1CDA37F8-9EF1-4349-8473-E529DC09482D}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Forum.Pages", "src\Forum.Pages\Forum.Pages.csproj", "{C8B01E88-BD47-4CA9-8A20-F93F02AC9EB3}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Forum.Tests.Library", "test\Forum.Tests.Library\Forum.Tests.Library.csproj", "{A4620D85-A3D9-4EEB-ADF2-60ADAE7BD779}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -35,11 +37,16 @@ Global
 		{C8B01E88-BD47-4CA9-8A20-F93F02AC9EB3}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C8B01E88-BD47-4CA9-8A20-F93F02AC9EB3}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{C8B01E88-BD47-4CA9-8A20-F93F02AC9EB3}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A4620D85-A3D9-4EEB-ADF2-60ADAE7BD779}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A4620D85-A3D9-4EEB-ADF2-60ADAE7BD779}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A4620D85-A3D9-4EEB-ADF2-60ADAE7BD779}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A4620D85-A3D9-4EEB-ADF2-60ADAE7BD779}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{440F16AA-0F6A-4467-ACB1-084B74DF0524} = {BB4873FD-8608-4D82-A597-405F86178DB9}
 		{CAACC8DF-000F-4486-B38D-936A63C7999D} = {BB4873FD-8608-4D82-A597-405F86178DB9}
 		{1CDA37F8-9EF1-4349-8473-E529DC09482D} = {E96FD598-6925-4A2B-B1A1-5B03AAB9C1B7}
 		{C8B01E88-BD47-4CA9-8A20-F93F02AC9EB3} = {BB4873FD-8608-4D82-A597-405F86178DB9}
+		{A4620D85-A3D9-4EEB-ADF2-60ADAE7BD779} = {E96FD598-6925-4A2B-B1A1-5B03AAB9C1B7}
 	EndGlobalSection
 EndGlobal
