+++ src/Forum.Data/Database.cs
@@ -19,30 +19,4 @@ public class Database : IDatabase
     {
         return new NpgsqlConnection(_connectionString);
     }
-
-    public void Initialize()
-    {
-        DropTopicsTable();
-        CreateTopicsTable();
-    }
-
-    private async void DropTopicsTable()
-    {
-        using var connection = Connect();
-        const string sql = @"drop table if exists topics;";
-
-        await connection.ExecuteAsync(sql);
-    }
-    
-    private async void CreateTopicsTable()
-    {
-        using var connection = Connect();
-        const string sql = @"create table topics (
-                                id          serial primary key,
-                                name        text   not null unique,
-                                description text
-                            );";
-
-        await connection.ExecuteAsync(sql);
-    }
 }
