+++ src/Forum.Data/Interfaces/IDatabase.cs
@@ -1,5 +1,4 @@
 ﻿using System.Data;
-using Dapper;
 
 namespace Forum.Data.Interfaces;
 
@@ -10,10 +9,4 @@ public interface IDatabase
     /// </summary>
     /// <returns>An open connection to the application database.</returns>
     public IDbConnection Connect();
-
-    /// <summary>
-    /// Initialize the database by creating tables needed for the application database, ensuring to also drop these
-    /// tables if they already exist.  
-    /// </summary>
-    public void Initialize();
 }
